OPTIMIZE: Changed wall seg rendering. Previously we were calculating whether each wall seg, in all visible subsectors faced the viewer at least three times per seg per frame unnecessarily. This calculation is now done once per frame per seg in visible subsectors and seg_t->flags updated with SEGINF_FACINGFRONT if true. Cleaned up con_bind.c, event bindings are now put into separate arrays depending on event_t->type. This plus the other recent changes enabled for the current code to be simplified greatly. Fixed numerous compiler warnings pointed out by GCC (thanks Yagisan). General cleanup.

+++ doomsday/engine/portable/src/sys_sock.c
@@ -17,7 +17,7 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, 
+ * Foundation, Inc., 51 Franklin St, Fifth Floor,
  * Boston, MA  02110-1301  USA
  */
 
@@ -94,7 +94,7 @@ void N_SockShutdown(void)
 void N_SockPrintf(socket_t s, const char *format, ...)
 {
     char    buf[512];
-    int     length;
+    size_t  length;
     va_list args;
 
     // Print the message into the buffer.
